requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
colorthief==0.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
paramiko==2.7.2
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
sshtunnel==0.4.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.0
whitenoise==5.2.0
wrapt==1.12.1
